for all snapshots change sid type to long

+++ hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/protocol/Block.java
@@ -179,7 +179,7 @@ public class Block implements Writable, Comparable<Block> {
     return sid;
   }
   
-  public void setSid(int id) {
+  public void setSid(long id) {
     sid = id;
   }
